split translatable to three files by class
